Added container; successfully require() modules
The TF popup has a container that loads the main interface in an
iframe.  This is the beginning of that structure in brunch-test.
container.{js,html} is the container, and {deps,main_tl}.js are
the main interface.

Also includes changes so that the various modules can be
require()d successfully.